+++ trunk/Code/Subasta.DomainServices/Game/ICandidatePlayer.cs
@@ -5,6 +5,6 @@ namespace Subasta.DomainServices.Game
 {
 	public interface ICandidatePlayer
 	{
-		IExplorationStatus PlayCandidate(IExplorationStatus currentStatus, int playerPosition, ICard candidate);//TODO: REMOVE PLAYER POSITION AS IS ALWAYS THE TURN
+		IExplorationStatus PlayCandidate(IExplorationStatus currentStatus, byte playerPosition, ICard candidate);//TODO: REMOVE PLAYER POSITION AS IS ALWAYS THE TURN
 	}
 }
